+++ packages/openapi-python/src/index.ts
@@ -76,17 +76,21 @@ import type { UserConfig } from '~/config/types';
 
 colors.enabled = colorSupport().hasBasic;
 
-// export { createClient } from '~/generate';
+export { createClient } from '~/generate';
 
 /**
- * Type helper for openapi-ts.config.ts, returns {@link MaybeArray<UserConfig>} object(s)
+ * Type helper for configuration object, returns {@link MaybeArray<UserConfig>} object(s)
  */
-export const defineConfig = async <T extends MaybeArray<UserConfig>>(
+export async function defineConfig<T extends MaybeArray<UserConfig>>(
   config: LazyOrAsync<T>,
-): Promise<T> => (typeof config === 'function' ? await config() : config);
+): Promise<T> {
+  return typeof config === 'function' ? await config() : config;
+}
 
+export { Logger } from '@hey-api/codegen-core';
 // export { defaultPaginationKeywords } from '~/config/parser';
 // export { defaultPlugins } from '~/config/plugins';
+export type { UserConfig } from '~/config/types';
 // export type { IR } from '~/ir/types';
 // export { OperationPath, OperationStrategy } from '~/openApi/shared/locations';
 // export type {
@@ -99,8 +103,6 @@ export const defineConfig = async <T extends MaybeArray<UserConfig>>(
 //   OpenApiSchemaObject,
 // } from '~/openApi/types';
 // export type { DefinePlugin, Plugin } from '~/plugins';
-// export type { AngularClient } from '~/plugins/@hey-api/client-angular';
-// export type { AxiosClient } from '~/plugins/@hey-api/client-axios';
 // export {
 //   clientDefaultConfig,
 //   clientDefaultMeta,
@@ -108,13 +110,6 @@ export const defineConfig = async <T extends MaybeArray<UserConfig>>(
 // export { clientPluginHandler } from '~/plugins/@hey-api/client-core/plugin';
 // export type { Client } from '~/plugins/@hey-api/client-core/types';
 // export type { FetchClient } from '~/plugins/@hey-api/client-fetch';
-// export type { NextClient } from '~/plugins/@hey-api/client-next';
-// export type { NuxtClient } from '~/plugins/@hey-api/client-nuxt';
-// export type { OfetchClient } from '~/plugins/@hey-api/client-ofetch';
-// export type { ExpressionTransformer } from '~/plugins/@hey-api/transformers/expressions';
-// export type { TypeTransformer } from '~/plugins/@hey-api/transformers/types';
 // export { definePluginConfig } from '~/plugins/shared/utils/config';
 // export * from '~/ts-dsl';
-// export type { UserConfig } from '~/types/config';
 // export { utils } from '~/utils/exports';
-// export { Logger } from '~/utils/logger';