+++ packages/openapi-ts/src/index.ts
@@ -72,21 +72,25 @@ import colors from 'ansi-colors';
 // @ts-expect-error
 import colorSupport from 'color-support';
 
-import type { UserConfig } from '~/types/config';
+import type { UserConfig } from '~/config/types';
 
 colors.enabled = colorSupport().hasBasic;
 
 export { createClient } from '~/generate';
 
 /**
- * Type helper for openapi-ts.config.ts, returns {@link MaybeArray<UserConfig>} object(s)
+ * Type helper for configuration object, returns {@link MaybeArray<UserConfig>} object(s)
  */
-export const defineConfig = async <T extends MaybeArray<UserConfig>>(
+export async function defineConfig<T extends MaybeArray<UserConfig>>(
   config: LazyOrAsync<T>,
-): Promise<T> => (typeof config === 'function' ? await config() : config);
+): Promise<T> {
+  return typeof config === 'function' ? await config() : config;
+}
 
+export { Logger } from '@hey-api/codegen-core';
 export { defaultPaginationKeywords } from '~/config/parser';
 export { defaultPlugins } from '~/config/plugins';
+export type { UserConfig } from '~/config/types';
 export type { IR } from '~/ir/types';
 export { OperationPath, OperationStrategy } from '~/openApi/shared/locations';
 export type {
@@ -115,6 +119,4 @@ export type { ExpressionTransformer } from '~/plugins/@hey-api/transformers/expr
 export type { TypeTransformer } from '~/plugins/@hey-api/transformers/types';
 export { definePluginConfig } from '~/plugins/shared/utils/config';
 export * from '~/ts-dsl';
-export type { UserConfig } from '~/types/config';
 export { utils } from '~/utils/exports';
-export { Logger } from '~/utils/logger';